requirements.txt
discord.py
wolframalpha
mcstatus
psycopg2
PyEnchant
PyDictionary
requests
